Feature/auction

+++ carrot/app/product/services.py
@@ -6,64 +6,76 @@ from carrot.app.user.schemas import UserOnboardingRequest, UserUpdateRequest
 from carrot.app.product.models import Product, UserProduct
 from carrot.app.product.repositories import ProductRepository
 from carrot.app.product.exceptions import NotYourProductException, InvalidProductIDException
+from carrot.app.image.services import ImageService
 from carrot.common.exceptions import InvalidFormatException
-# from carrot.app.image.models import ProductImage
 
 class ProductService:
-    def __init__(self, product_repository: Annotated[ProductRepository, Depends()]) -> None:
+    def __init__(self, product_repository: Annotated[ProductRepository, Depends()], image_service: Annotated[ImageService, Depends()]) -> None:
         self.repository = product_repository
+        self.image_service = image_service
 
-    async def create_post(self, user_id: str, title: str, content: str, price: int, category_id: str) -> Product:
-        # image_objects = [ProductImage(image_url=img_url) for img_url in images]
+    async def create_post(self, user_id: str, title: str, image_ids: list, content: str, price: int, category_id: str, region_id: str) -> Product:
         product = Product(
             owner_id = user_id,
             title = title,
-            # images = image_objects,
+            image_ids = image_ids,
             content = content,
             price = price,
             category_id = category_id,
+            region_id = region_id,
         )
         
         new = await self.repository.create_post(product)
         return new
 
-    async def update_post(self, user_id: str, id: str, title: str, content: str, price: int, category_id: str) -> Product:
-        product = await self.repository.get_post_by_id(id)
+    async def update_post(self, user_id: str, id: str, title: str, image_ids: list, content: str, price: int, category_id: str, region_id: str) -> Product:
+        product = await self.repository.get_post_by_product_id(id)
 
         if product is None:
             raise InvalidProductIDException
         
         if user_id != product.owner_id:
             raise NotYourProductException
-        if title is not None:
-            product.title = title
-        # if images is not None:
-        #     image_objects = [ProductImage(image_url=img_url) for img_url in images]
-        #     product.images = image_objects
-        if content is not None:
-            product.content = content
-        if price is not None:
-            product.price = price
-        if category_id is not None:
-            product.category_id = category_id
+        
+        product.title = title
+        product.image_ids = image_ids
+        product.content = content
+        product.price = price
+        product.category_id = category_id
+        product.region_id = region_id
         
         updated = await self.repository.update_post(product)
         return updated
     
-    async def view_post_my(self, user_id: str):
-        products = await self.repository.get_post_by_user_id(user_id)
+    async def view_post_by_product_id(self, product_id: str):
+        product = await self.repository.get_post_by_product_id(product_id)
         
-        return products
-        
-    async def view_post_all(self):
-        products = await self.repository.get_post_all()
-        
-        return products
-    
-    async def remove_post(self, user_id:str, id: str):
-        product = await self.repository.get_post_by_id(id)
         if product is None:
             raise InvalidProductIDException
         
-        await self.repository.remove_post(product)
         return product
+    
+    async def view_posts_by_query(self, user_id: str | None, keyword: str | None, region_id: str | None):
+        products = await self.repository.get_posts_by_query(user_id, keyword, region_id)
+        
+        return products
+    
+    async def view_posts_all(self):
+        products = await self.repository.get_posts_all()
+        
+        return products
+    
+    async def remove_post(self, user_id: str, product_id: str):
+        product = await self.repository.get_post_by_product_id(product_id)
+        
+        if product is None:
+            raise InvalidProductIDException
+        
+        if user_id != product.owner_id:
+            raise NotYourProductException
+        
+        image_ids = product.image_ids
+        for id in image_ids:
+            await self.image_service.remove_product_image(id)
+            
+        await self.repository.remove_post(product)
